Add types balance module
